pyproject.toml
[project]
name = "grabpic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.128.6",
    "firebase-admin>=7.1.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary==2.9.6",
    "pwdlib[argon2]==0.3.0",
    "pydantic==2.10.4",
    "pyjwt>=2.11.0",
    "python-dotenv==1.2.1",
    "sqlalchemy==2.0.19",
    "uvicorn==0.34.0",
    "insightface>=0.7.3",
    "numpy==2.3.0",
    "onnxruntime-gpu>=1.24.1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "retina-face>=0.0.17",
    "tf-keras>=2.20.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "gunicorn==25.1.0",
    "argon2-cffi>=25.1.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
